backend/requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
python-dotenv
PyJWT
gunicorn
psycopg2-binary